Styling tweaks for large type or small screens, gets rid of start new project component
